source/servicenow_function/requirements.txt
google-cloud-logging
google-cloud-secret-manager
google-cloud-storage
google-cloud-datastore
google-auth
requests